requirements.txt
# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1

# LangChain Ecosystem + LangGraph
# langchain==0.2.1
langchain-openai==0.1.8
langchain-community==0.2.1
langchain-chroma==0.1.1
langchain-text-splitters==0.2.0
langchain-groq==0.1.6
langgraph==0.1.19
langchain>=0.2.5


# LLM Providers
openai==1.30.1
groq==0.9.0

# Vector DB
chromadb==0.5.0

# Local Embeddings (FREE -- runs on CPU, no API key needed)
sentence-transformers==2.7.0
torch==2.3.0
transformers==4.41.1
huggingface-hub==0.23.2

# HTTP & APIs
httpx==0.27.0
requests==2.31.0
arxiv==2.1.0

# PDF Processing & Export
pypdf==4.2.0
pdfplumber==0.11.0
markdown2==2.4.13
reportlab==4.2.0

# Frontend
streamlit==1.35.0

# Utilities
tenacity==8.3.0
loguru==0.7.2
python-multipart==0.0.9
aiofiles==23.2.1
numpy==1.26.4